+++ packages/toolkit/src/index.ts
@@ -1,4 +1,5 @@
 export { Toolkit } from './backend/toolkit';
+export type { ToolkitConnection } from './backend/toolkit';
 export type { ToolkitOptions } from './backend/options';
 
 export type { AnyComponent } from './backend/components/base';